requirements.txt
# Polymarket Spike Hunter Bot - Dependencies
# ============================================

# Polymarket API client
py-clob-client>=0.17.0

# Async HTTP client
aiohttp>=3.9.0

# Configuration and validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0

# Notifications
python-telegram-bot>=20.0
discord-webhook>=1.3.0

# Logging
structlog>=23.2.0

# Environment variables
python-dotenv>=1.0.0

# Data processing
pandas>=2.1.0
numpy>=1.26.0

# Dashboard
streamlit>=1.29.0
plotly>=5.18.0

# Web3 for Polymarket
web3>=6.11.0
eth-account>=0.10.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
isort>=5.12.0
mypy>=1.7.0